+++ trunk/libmesh/include/cell_inf_prism12.h
@@ -1,4 +1,4 @@
-// $Id: cell_inf_prism12.h,v 1.14 2003-08-07 19:25:30 ddreyer Exp $
+// $Id: cell_inf_prism12.h,v 1.15 2003-08-18 14:44:51 ddreyer Exp $
 
 // The Next Great Finite Element Library.
 // Copyright (C) 2002  Benjamin S. Kirk, John W. Peterson
@@ -116,8 +116,8 @@ public:
    * that defines the \f$ n^{th} \f$ second-order node.
    * Note that \p n is counted as depicted above, \f$ 6 \le n < 12 \f$.
    */
-  unsigned int second_order_adjacent_vertex (const unsigned int n,
-					     const unsigned int v) const;
+  unsigned short int second_order_adjacent_vertex (const unsigned int n,
+						   const unsigned int v) const;
 
 
 
@@ -148,7 +148,7 @@ private:
    * Matrix that tells which vertices define the location
    * of mid-side (or second-order) nodes
    */
-  static const unsigned int _second_order_adjacent_vertices[6][2];
+  static const unsigned short int _second_order_adjacent_vertices[6][2];
   
 };
